Make decomposer yield spans
The purpose of the decomposer is to indicate sections of the input. I think spans are more suited for this than strings.
Also, it is (maybe subjectively) easier to convert from a span to a string than vice versa.
Also, spans are smaller.
Also, there is no runtime guarantee that the strings previously yielded by the decomposer were actually parts of the original input.