Rename `Span::from_slice` to clarify functionality
This is to make it clear that the function will not determine the location of a subslice within a slice, which in theory should be obvious but in practice may not be, and we're only human after all.
`entire_slice` is a better name that makes it clear that the span will refer to the whole of the slice.