Refactor decomposer and finish lexer
I would have liked to have separated this into two commits, but the changes were far too intertwined.

First, I refactored the decomposer heavily, first making it into a custom iterator, and then making some internal changes to avoid blowing the stack (I added the `all_commas` test to ensure this).
The main upshot of reimplementing the decomposer as a custom iterator, besides fixing bugs and removing the use of the three nightly features, is the `next_no_decompose` method, which allows the lexer to not decompose when looking for the ending delimiter of a delimited quoted. It acts similarly to just calling `next` on the internal iterator, but makes sure, if the decomposer was in the middle of decomposing a chunk, to first yield the remaining part of that chunk.

Then, I modified the lexer to make use of that method, which the delimited quote parsing code a bit simpler, and also more correct. Before, lexing `"zoi gy. text .gy"` would yield `["zoi", "gy", ". text .", "gy"]`, whereas now it yields `["zoi", "gy", "text", "gy"]`. It acts the same as if the pauses were stripped from the text portion, but accomplishes that behavior in a more reliable way.