Assert that some rules consumed all input
If we want to make sure that a given string is a gismu, we don't only want to check that it starts with a gismu, but that the gismu rule matched the entire string with none left over.
The `succeeded_and_consumed_all` method provides that functionality.
Use it wherever it is more semantically correct.