Start working on parser
I created a new branch for the parser so I can commit as I implement aspects of the language.

Since this is the first commit in that branch, I will describe my new approach to writing the parser here: Rather than copying rules from Ilmentufa, I'm going to read the CLL and implement features as the CLL mentions them. Thus most of my commits in the future in this branch will be something like "Catch up to CLL 5.6" (which is in fact where I am currently, approximately).