Fix vau termination
`vau` is a bit tricky since the exact formation of the bridi before it determines how many `vau` will be accepted. For example if there is a GIhA used then that will allow an extra `vau`.

Previously we used `Separated` and then `Option<Vau>`. `Separated` uses `Box<[(Separator, Item)]>` and `many0`. Instead, put the `Box` in a new type with `Option<Vau>` and parse the box with `many1`, then put an `Option` of *that* type into the contexts where we had `Separated` and `Option<Vau>` before.